Send Activation email from scheduler
When an order is activated via the admin page, a correspondent email is
sent out to the user, containing the license key (module
Servicelicense). However, this is not happening when we pay via Paypal
and the order gets automatically activated on the next scheduler run.

The reason being is that the $di['loggedin_admin'] call throws an
exception and the email code does not get executed. In reality we do not
need to collect sensitive data, all we need is the client_id which is
already present in the $order obejct, therefore we can send out the
email.

This is very important for automatic shops selling license. Although the
license can be viewed on the user's Services page, anyone new to BB will
interpret as having paid and not received the correspondent key.

See also

http://www.boxbilling.com/forum/boxbilling-support/account-activated-mail-problems-please-help